修改一个错误

+++ app/controllers/api/v1/maps_controller.rb
@@ -49,24 +49,24 @@ class Api::V1::MapsController < Api::V1::ApplicationController
   private
 
   def validate_client_state
-    result = true
     ip = request.remote_ip
-    result = false if ip.blank?
+    return false if ip.blank?
     past_ip = IpAddress.find{|i| i.ip == ip}
     if past_ip.blank?
       IpAddress.create ip: ip
+      return true
     else
       if Time.now() - past_ip.created_at > 3600
         past_ip.destroy
         IpAddress.create ip: ip
+        return true
       elsif past_ip.counter > 100
-        result = false
+        return false
       else
         past_ip.counter += 1
         past_ip.save
+        return true
       end
     end
-    result
   end
-  
 end
